requirements.txt
Django
django-debug-toolbar==2.2
python-decouple==3.3
whitenoise
djangorestframework
